% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_511_metadata}
\alias{get_511_metadata}
\title{Get metadata about MTC 511 operators}
\usage{
get_511_metadata()
}
\value{
a tibble of operator names, privatecodes, and primary modes
}
\description{
Get metadata about MTC 511 operators
}
\examples{
\dontrun{
#set your api key as an environmental variable
Sys.setenv(APIKEY511="yourkeyhere")
#if you don't have a key, you can get one here:
#https://511.org/developers/list/tokens/create
operator_df <- get_511_metadata()
head(operator_df)
}
}
