% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_511_url}
\alias{get_511_url}
\title{Use the trread import_gtfs function to import an MTC 511 GTFS zip file into a list of R dataframes}
\usage{
get_511_url(privatecode)
}
\arguments{
\item{privatecode}{this is the shortcode used by 511 to refer to operators. you can find these on the tibble returned from get_511_metadata}
}
\value{
a list of GTFS dataframes
}
\description{
Use the trread import_gtfs function to import an MTC 511 GTFS zip file into a list of R dataframes
}
\examples{
\dontrun{
#set your api key as an environmental variable
Sys.setenv(APIKEY511="yourkeyhere")
#if you don't have a key, you can get one here:
#https://511.org/developers/list/tokens/create
operator_df <- xml511_to_tibble()
bart_code <- df1[df1$name=='Bay Area Rapid Transit',]$privatecode
bart_gtfs_data <- get_mtc_511_gtfs(bart_code)
}
}
