% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{xml511_to_tibble}
\alias{xml511_to_tibble}
\title{parse 511 xml into a table for each operator}
\usage{
xml511_to_tibble(oneop)
}
\value{
a tibble of operator names, privatecodes, and primary modes
}
\description{
parse 511 xml into a table for each operator
}
\keyword{internal}
